requirements.txt
Django~=4.0.4
djangorestframework~=3.13.1
pytz~=2022.1
channels~=3.0.5
asgiref~=3.5.0